+++ frontend/src/pages/Notifications/NotificationListItems.js
@@ -8,12 +8,13 @@ import ListItemText from "@material-ui/core/ListItemText";
 import Unread from "@material-ui/icons/Email";
 import Read from "@material-ui/icons/MailOutline";
 import LaunchIcon from "@material-ui/icons/ZoomIn";
-import dayjs from "dayjs";
 import React from "react";
 import classNames from "classnames";
 
 import { intentMapping, parseURI, getParentData, isAllowedToSee } from "./helper";
 import strings from "../../localizeStrings";
+import { dateFormat } from "../../helper";
+import dayjs from "dayjs";
 const styles = theme => ({
   row: {
     display: "flex",
@@ -58,12 +59,13 @@ const NotificationListItems = ({
   return notifications.map((notification, index) => {
     const message = intentMapping(notification);
     const { businessEvent, id, isRead, metadata } = notification;
-    const createdAt = dayjs(businessEvent.time).fromNow();
+    const createdAt = dayjs(businessEvent.time).format(dateFormat());
     const redirectUri = parseURI({
       projectId: metadata.project ? metadata.project.id : undefined,
       subprojectId: metadata.subproject ? metadata.subproject.id : undefined
     });
     const testLabel = `notification-${isRead ? "read" : "unread"}`;
+    const dateTestLabel = "dateOfNotification";
     const { projectDisplayName, subprojectDisplayName } = getParentData(notification);
     return (
       <div key={index}>
@@ -92,6 +94,7 @@ const NotificationListItems = ({
             primary={message}
           />
           <ListItemText
+            data-test={`${dateTestLabel}-${index}`}
             className={classes.author}
             component="div"
             primary={businessEvent.publisher}